// examples/start-full-stack/src/generated/api/schema.ts
export const petCategorySchema = z.enum(["dog", "cat", "bird", "fish", "reptile"])
export const petStatusSchema = z.enum(["available", "pending", "sold"])
// ...
export const updatePetInputSchema = z.object({
  name: z.string().nullish(),
  category: petCategorySchema.nullish(),
  status: petStatusSchema.nullish(),
  tags: z.array(z.string()).nullish(),
  photoUrl: z.url().nullish()
})
// ...
export const updatePetRequestSchema = updatePetInputSchema
// ...
export type PetCategory = z.infer<typeof petCategorySchema>
export type PetStatus = z.infer<typeof petStatusSchema>
// ...
export type UpdatePetRequest = z.infer<typeof updatePetRequestSchema>